+++ packages/std/src/testing/mock.rs
@@ -24,7 +24,9 @@ use crate::ibc::{
     IbcTimeoutBlock,
 };
 #[cfg(feature = "ibc2")]
-use crate::ibc2::{Ibc2PacketAckMsg, Ibc2PacketReceiveMsg, Ibc2PacketTimeoutMsg, Ibc2Payload};
+use crate::ibc2::{
+    Ibc2PacketAckMsg, Ibc2PacketReceiveMsg, Ibc2PacketSendMsg, Ibc2PacketTimeoutMsg, Ibc2Payload,
+};
 #[cfg(feature = "cosmwasm_1_1")]
 use crate::query::SupplyResponse;
 #[cfg(feature = "staking")]
@@ -714,6 +716,25 @@ pub fn mock_ibc2_packet_timeout(data: &impl Serialize) -> StdResult<Ibc2PacketTi
         "destination_client".to_string(),
         1,
         Addr::unchecked("relayer"),
+    ))
+}
+
+/// Creates a Ibc2PacketTimeoutMsg for testing ibc2_packet_timeout.
+#[cfg(feature = "ibc2")]
+pub fn mock_ibc2_packet_send(data: &impl Serialize) -> StdResult<Ibc2PacketSendMsg> {
+    let payload = Ibc2Payload {
+        source_port: "wasm2srcport".to_string(),
+        destination_port: "wasm2destport".to_string(),
+        version: "v2".to_string(),
+        encoding: "json".to_string(),
+        value: to_json_binary(data)?,
+    };
+    Ok(Ibc2PacketSendMsg::new(
+        payload,
+        "source_client".to_string(),
+        "destination_client".to_string(),
+        1,
+        Addr::unchecked("signer_contract"),
     ))
 }