+++ packages/vm/src/calls.rs
@@ -15,14 +15,15 @@ use cosmwasm_std::{
     IbcPacketAckMsg, IbcPacketReceiveMsg, IbcPacketTimeoutMsg,
 };
 
-#[cfg(feature = "ibc2")]
-use cosmwasm_std::{Ibc2PacketAckMsg, Ibc2PacketReceiveMsg, Ibc2PacketTimeoutMsg};
-
 use crate::backend::{BackendApi, Querier, Storage};
 use crate::conversion::ref_to_u32;
 use crate::errors::{VmError, VmResult};
 use crate::instance::Instance;
 use crate::serde::{from_slice, to_vec};
+#[cfg(feature = "ibc2")]
+use cosmwasm_std::{
+    Ibc2PacketAckMsg, Ibc2PacketReceiveMsg, Ibc2PacketSendMsg, Ibc2PacketTimeoutMsg,
+};
 
 /// The limits in here protect the host from allocating an unreasonable amount of memory
 /// and copying an unreasonable amount of data.
@@ -67,6 +68,8 @@ mod read_limits {
     pub const RESULT_IBC_SOURCE_CALLBACK: usize = 64 * MI;
     /// Max length (in bytes) of the result data from a ibc_destination_callback call.
     pub const RESULT_IBC_DESTINATION_CALLBACK: usize = 64 * MI;
+    #[cfg(feature = "ibc2")]
+    pub const RESULT_IBC2_PACKET_SEND: usize = 64 * MI;
 }
 
 /// The limits for the JSON deserialization.
@@ -110,6 +113,9 @@ mod deserialization_limits {
     pub const RESULT_IBC_SOURCE_CALLBACK: usize = 256 * KI;
     /// Max length (in bytes) of the result data from a ibc_destination_callback call.
     pub const RESULT_IBC_DESTINATION_CALLBACK: usize = 256 * KI;
+    /// Max length (in bytes) of the result data from a ibc_packet_receive call.
+    #[cfg(feature = "ibc2")]
+    pub const RESULT_IBC2_PACKET_SEND: usize = 256 * KI;
 }
 
 pub fn call_instantiate<A, S, Q, U>(
@@ -775,6 +781,45 @@ where
     )
 }
 
+#[cfg(feature = "ibc2")]
+pub fn call_ibc2_packet_send_raw<A, S, Q>(
+    instance: &mut Instance<A, S, Q>,
+    env: &[u8],
+    msg: &[u8],
+) -> VmResult<Vec<u8>>
+where
+    A: BackendApi + 'static,
+    S: Storage + 'static,
+    Q: Querier + 'static,
+{
+    instance.set_storage_readonly(false);
+    call_raw(
+        instance,
+        "ibc2_packet_send",
+        &[env, msg],
+        read_limits::RESULT_IBC2_PACKET_SEND,
+    )
+}
+
+#[cfg(feature = "ibc2")]
+pub fn call_ibc2_packet_send<A, S, Q, U>(
+    instance: &mut Instance<A, S, Q>,
+    env: &Env,
+    msg: &Ibc2PacketSendMsg,
+) -> VmResult<ContractResult<IbcBasicResponse<U>>>
+where
+    A: BackendApi + 'static,
+    S: Storage + 'static,
+    Q: Querier + 'static,
+    U: DeserializeOwned + CustomMsg,
+{
+    let env = to_vec(env)?;
+    let msg = to_vec(msg)?;
+    let data = call_ibc2_packet_send_raw(instance, &env, &msg)?;
+    let result = from_slice(&data, deserialization_limits::RESULT_IBC2_PACKET_SEND)?;
+    Ok(result)
+}
+
 pub fn call_ibc_source_callback_raw<A, S, Q>(
     instance: &mut Instance<A, S, Q>,
     env: &[u8],
@@ -1362,7 +1407,8 @@ mod tests {
     mod ibc2 {
         use super::*;
         use cosmwasm_std::testing::{
-            mock_ibc2_packet_ack, mock_ibc2_packet_recv, mock_ibc2_packet_timeout,
+            mock_ibc2_packet_ack, mock_ibc2_packet_recv, mock_ibc2_packet_send,
+            mock_ibc2_packet_timeout,
         };
         static IBC2: &[u8] = include_bytes!("../testdata/ibc2.wasm");
 
@@ -1425,6 +1471,26 @@ mod tests {
             let ibc2_msg = br#"SomeRandomMsg"#;
             let ibc2_msg = mock_ibc2_packet_timeout(ibc2_msg).unwrap();
             call_ibc2_packet_timeout::<_, _, _, Empty>(&mut instance, &mock_env(), &ibc2_msg)
+                .unwrap()
+                .unwrap();
+        }
+
+        #[test]
+        fn call_ibc2_packet_send_works() {
+            // init
+            let mut instance = mock_instance(IBC2, &[]);
+            let info = mock_info("creator", &[]);
+            let instantiate_msg = br#"{}"#;
+            call_instantiate::<_, _, _, Empty>(&mut instance, &mock_env(), &info, instantiate_msg)
+                .unwrap()
+                .unwrap();
+
+            let ibc2_msg = IbcPayload {
+                response_without_ack: false,
+                send_async_ack_for_prev_msg: false,
+            };
+            let ibc2_sent = mock_ibc2_packet_send(&ibc2_msg).unwrap();
+            call_ibc2_packet_send::<_, _, _, Empty>(&mut instance, &mock_env(), &ibc2_sent)
                 .unwrap()
                 .unwrap();
         }