+++ packages/vm/src/lib.rs
@@ -35,7 +35,8 @@ pub use crate::calls::{
 #[cfg(feature = "ibc2")]
 pub use crate::calls::{
     call_ibc2_packet_ack, call_ibc2_packet_ack_raw, call_ibc2_packet_receive,
-    call_ibc2_packet_receive_raw, call_ibc2_packet_timeout, call_ibc2_packet_timeout_raw,
+    call_ibc2_packet_receive_raw, call_ibc2_packet_send, call_ibc2_packet_send_raw,
+    call_ibc2_packet_timeout, call_ibc2_packet_timeout_raw,
 };
 #[cfg(feature = "stargate")]
 pub use crate::calls::{